requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
elasticsearch
ollama
jinja2
pandas
openpyxl
streamlit
pydantic
python-dotenv
requests
